Finished elite mob challenges, misc fixes

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/custom/ch/mob/city/SeniorMonkH.java
@@ -5,7 +5,9 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Blindness;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
+import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Corruption;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Cripple;
+import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Roots;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Vertigo;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.Monk;
@@ -17,6 +19,10 @@ import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.KindOfWeapon;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.Pasty;
+import com.shatteredpixel.shatteredpixeldungeon.items.rings.Ring;
+import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfAccuracy;
+import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfEvasion;
+import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfWealth;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.Gauntlet;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.Gloves;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.SeniorSprite;
@@ -32,6 +38,9 @@ public class SeniorMonkH extends Monk {
         lootChance=1f;
         loot = new Pasty();
     }
+    {
+        immunities.add(Corruption.class);
+    }
 
     @Override
     public float attackDelay(){return super.attackDelay()*0.85f;}
@@ -44,29 +53,36 @@ public class SeniorMonkH extends Monk {
         if(enemy == Dungeon.hero && hitsForSkill>6){
             hitsForSkill = 0;
             Hero hero = Dungeon.hero;
-            BallisticaFloat ba = new BallisticaFloat(enemy.pos, GME.angle(pos, enemy.pos)+ Random.Float(-22.5f, 22.5f), 8, BallisticaFloat.PROJECTILE);
-            Actor.add(new Pushing(hero, hero.pos, ba.collisionPosI, new Callback() {
-                @Override
-                public void call() {
-                    Camera.main.shake(3f, 0.5f);
-                    Buff.affect(hero, Blindness.class, 6f);
-                    Buff.affect(hero, Vertigo.class, 8f);
-                    Buff.affect(hero, Cripple.class, 8f);
-                    hero.sprite.centerEmitter().burst(Speck.factory(Speck.STAR), 10);
-                    hero.pos = ba.collisionPosI;
-                    Dungeon.level.occupyCell(hero);
-                }
-            }));
-            KindOfWeapon wep = hero.belongings.weapon;
-            if(wep != null && !wep.cursed && !(wep instanceof Gloves || wep instanceof Gauntlet)){
-                Dungeon.level.drop(wep, hero.pos).sprite.drop();
-                hero.belongings.weapon = null;
-                Dungeon.quickslot.convertToPlaceholder(wep);
-                Item.updateQuickslot();
-            }
+            //cannot hitback rooted enemy!
+            if(hero.buff(Roots.class)==null) {
+                BallisticaFloat ba = new BallisticaFloat(enemy.pos, GME.angle(pos, enemy.pos) + Random.Float(-22.5f, 22.5f), 8, BallisticaFloat.PROJECTILE);
+                Actor.add(new Pushing(hero, hero.pos, ba.collisionPosI, new Callback() {
+                    @Override
+                    public void call() {
+                        Camera.main.shake(3f, 0.5f);
+                        Buff.affect(hero, Blindness.class, 6f);
+                        Buff.affect(hero, Vertigo.class, 8f);
+                        Buff.affect(hero, Cripple.class, 8f);
+                        hero.spendAndNext(TICK);
 
-            GLog.n(M.L(this, "hard_strike"));
-            spend(attackDelay());
+                        hero.sprite.centerEmitter().burst(Speck.factory(Speck.STAR), 10);
+                        hero.pos = ba.collisionPosI;
+                        Dungeon.level.occupyCell(hero);
+
+                    }
+                }));
+                KindOfWeapon wep = hero.belongings.weapon;
+                if (wep != null && !wep.cursed && !(wep instanceof Gloves || wep instanceof Gauntlet)) {
+                    Dungeon.level.drop(wep, hero.pos).sprite.drop();
+                    hero.belongings.weapon = null;
+                    Dungeon.quickslot.convertToPlaceholder(wep);
+                    Item.updateQuickslot();
+                }
+
+
+                GLog.n(M.L(this, "hard_strike"));
+                spend(attackDelay());
+            }
         }
         return super.attackProc(enemy, damage);
     }
@@ -90,5 +106,23 @@ public class SeniorMonkH extends Monk {
     public void restoreFromBundle(Bundle b){
         super.restoreFromBundle(b);
         hitsForSkill = b.getInt("hitsForSkill");
+    }
+
+    @Override
+    public void rollToDropLoot(){
+        if (Dungeon.hero.lvl <= maxLvl + 2){
+            float chance = 0.2f;
+            chance *= RingOfWealth.dropChanceMultiplier( Dungeon.hero );
+            chance = Math.min(0.4f, chance);
+            if(Random.Float()<chance){
+                Ring r;
+                r = Random.Int(2)==0? new RingOfAccuracy() : new RingOfEvasion();
+                r.level(Random.chances(new float[]{0.45f - 2f*chance, 0.3f + chance, 0.2f+chance/2f, 0.05f + chance/2f}));
+                Dungeon.level.drop(r, pos).sprite.drop();
+            }
+        }
+
+        super.rollToDropLoot();
+
     }
 }
